acces au gerant

+++ proxi-api/src/main/java/ma/jit/proxibanque/dao/GerantRepository.java
@@ -17,6 +17,6 @@ import ma.jit.proxibanque.entities.Gerant;
  *
  */
 @Repository
-public interface GerantRepository extends JpaRepository<Gerant, Integer>{
+public interface GerantRepository extends JpaRepository<Gerant, Long>{
 	
 }
